==> notebooks/model_creation/03_create_test_train.ipynb <==
# Create training and test set

**Goal**: Create test set and put it aside - this will only be used for benchmarking at the end. I will apply scikitlearn's stratifiedshufflesplit to try and sample evenly from different cell types, in case there are cell type specific signatures.

**Output**: Test dataset (used for model assessment only), training dataset (used for model training and hyperparameter selection)

**Assumptions (that could be tweaked)**:
- Normalization procedure is sufficient rescaling to facilitate effective model training (scanpy normalize_total + log1p)


# Import needed libraries
from sklearn.model_selection import StratifiedShuffleSplit
from sklearn.preprocessing import OneHotEncoder


import scanpy as sc
import numpy as np
import pandas as pd
import os
import pathlib

import matplotlib.cm as cm
import matplotlib.pyplot as plt

import pickle

from utils.config import *
from utils.analysis_variables import *

# Scanpy setup
sc.settings.verbosity = 3 # corresponds to hints

# Notebook setup
np.random.seed(15)

import warnings
warnings.filterwarnings('ignore')

# Important paths
notebook_name = "03_create_test_train"

# path_outdir_base = "../../output/20240221_import"
path_results = os.path.join(path_outdir_base, notebook_name)
os.makedirs(path_results, exist_ok=True)

path_input_data = os.path.join(path_outdir_base, "02_explore_data_cell_id", "adata_clusterlabels.h5ad")

# Import data and data preparation

adata = sc.read_h5ad(path_input_data)

# Create factor level that is target variable, Sex, plus the level to split on
adata.obs['Sex_Celltype'] = adata.obs.apply(lambda x: x.Sex + "_" + x.leiden_labeled, axis=1)

adata.obs.head(3)

# Create test dataset to put aside

Here, y is Sex_Celltype so that I can sample representatives from each Sex x Cell type combination.

# Define the number of splits and test size
n_splits = 1
test_size = 0.2

# Filter out the data with no known sex as an unlabeled test set
adata_unlabeled = adata[adata.obs['Sex'] == 'Unknown']
adata_labeled = adata[adata.obs['Sex'] != 'Unknown']


# Make X and y_label variables for these in array format
data = adata_labeled.X.toarray()

one_hot_encoder = OneHotEncoder(sparse=False, drop='first', handle_unknown='ignore')
target_sex_celltype = adata_labeled.obs.Sex_Celltype
target_sex = adata_labeled.obs.Sex
target_sex_celltype_encoded = one_hot_encoder.fit_transform(target_sex_celltype.astype('category').values.reshape(-1, 1))

stratified_splitter = StratifiedShuffleSplit(n_splits=n_splits, test_size=test_size, random_state=my_random_state)

for train_index, test_index in stratified_splitter.split(data, target_sex_celltype_encoded):
    # Get the training and test data
    X_train, X_test = data[train_index], data[test_index] # order of entries same as order in adata_labeled
    y_train, y_test = target_sex[train_index], target_sex[test_index] # Actually only include the one predictor we care about


print(f"Size of training data (N cell x gene): {X_train.shape}")
print(f"Size of test data: {X_test.shape}")

print("Training data N:")
adata_labeled.obs.iloc[train_index].value_counts(['Sex', 'leiden_labeled'])

print("Test data N:")
adata_labeled.obs.iloc[test_index].value_counts(['Sex', 'leiden_labeled'])

# Save data


with open(os.path.join(path_results, 'training_data.pkl'), 'wb') as f:
    pickle.dump({'X': X_train, 'Y': y_train}, f)

with open(os.path.join(path_results, 'test_data.pkl'), 'wb') as f:
    pickle.dump({'X': X_test, 'Y': y_test}, f)

adata_labeled.write(os.path.join(path_results, 'adata_labeled.h5ad'))
adata_unlabeled.write(os.path.join(path_results, 'adata_unlabeled.h5ad'))


path_results
