==> notebooks/model_creation/05a_model_selection.ipynb <==
# Model selection

**Goal**: Given a reasonable feature list, try assessing different models to figure out which model family might be most successful in fitting this data. The intent is to focus on high-level architectural decisions (e..g overall structure of model)

**Output**: A report with metrics to inform which model types are likely to perform better.

**TODO**:
- 


# Import needed libraries
from sklearn.model_selection import StratifiedShuffleSplit

import scanpy as sc
import numpy as np
import pandas as pd
import os
import pathlib

import matplotlib.cm as cm
import matplotlib.pyplot as plt

from utils.config import *
from utils.analysis_variables import *
from utils.analysis_functions import *

import pickle

from sklearn.feature_selection import SelectKBest, GenericUnivariateSelect, f_classif
from sklearn.model_selection import GridSearchCV
from sklearn.svm import SVC
from sklearn.pipeline import Pipeline
from sklearn.metrics import make_scorer, confusion_matrix


from sklearn.model_selection import GridSearchCV, train_test_split, RandomizedSearchCV
from sklearn.preprocessing import StandardScaler
from sklearn.ensemble import RandomForestClassifier, GradientBoostingClassifier
from sklearn.linear_model import LogisticRegression
from sklearn.svm import SVC
from sklearn.pipeline import Pipeline
from sklearn.datasets import make_classification
from sklearn.metrics import make_scorer, confusion_matrix

import pickle


# Scanpy setup
sc.settings.verbosity = 3 # corresponds to hints

# Notebook setup
np.random.seed(15)

import warnings
warnings.filterwarnings('ignore')

# Important paths
notebook_name = "05a_model_selection"

# path_outdir_base = "../../output/20240221_import"
path_results = os.path.join(path_outdir_base, notebook_name)
os.makedirs(path_results, exist_ok=True)

path_input_data = os.path.join(path_outdir_base, "03_create_test_train", "training_data.pkl")
path_input_adata = os.path.join(path_outdir_base, "03_create_test_train", "adata_labeled.h5ad")
path_input_features = os.path.join(path_outdir_base, "04_feature_selection", "dict_feature_data.pkl")

# Import data

with open(path_input_data, 'rb') as f:
    data_dict = pickle.load(f)

X_train = data_dict['X']
y_train = data_dict['Y']

adata_labeled = sc.read_h5ad(path_input_adata)

# open sexually dimorphic genes
with open(os.path.join(path_outdir_base, '01_data_import_and_filtering', 'all_sex_dimorph_genes.pkl'), 'rb') as file:
    all_sex_dimorph_genes = pickle.load(file)

adata_labeled.var

# Import possible features

with open(path_input_features, 'rb') as file:
    dict_feature_data = pickle.load(file)

# Add all sexually dimorphic gene as a geneset for comparison
adata_labeled.var['msigdb_all_sex_dimorph_genes'] = adata_labeled.var.index.map(lambda x: True if x in all_sex_dimorph_genes else False)
dict_feature_data.update({'Sexually Dimorphic Genes': adata_labeled.var.index[adata_labeled.var.msigdb_all_sex_dimorph_genes]})

# prelim_model_features = dict_feature_data["FPR (All)"]
# adata_model = adata_labeled[:, adata_labeled.var.index.isin(prelim_model_features)]

# Create a hyperparameter that is the feature set to use
X_train_featurefilter = {}

for feature_set in dict_feature_data.keys():
    mask = adata_labeled.var.index.isin(dict_feature_data[feature_set])

    X_train_featurefilter.update({feature_set: X_train[:,mask]} )
    

# # Subset training data to only include X_data corresponding to features we want to use
# mask_features = adata_labeled.var.index.isin(prelim_model_features)
# X_train_features = X_train[:, mask_features]

# print(f"Shape of X_train: {X_train_features.shape}, Shape of Y_train: {y_train.shape}")

# Pipeline + GridSearchCV to compare performance of different types of models

## Set up search

# Define scoring metrics

scoring = {
    'accuracy': 'accuracy',  # Default scoring for classification
    'specificity': make_scorer(specificity),
    'false_positive_rate': make_scorer(false_positive_rate, greater_is_better=False),  # Minimize false positive rate
    'false_negative_rate': make_scorer(false_negative_rate, greater_is_better=False),  # Minimize false negative rate; something funny abt this processing makes this neg
    'precision': make_scorer(precision)
}

# Define pipeline
pipeline = Pipeline([
    ('scaler', StandardScaler()),
    ('classifier', None), # classifier will be replaced during grid search
])


# Define hyperparameters to search
# param_grid = {
#     'classifier': [RandomForestClassifier(), SVC(), LogisticRegression(), GradientBoostingClassifier()],
#     'classifier__n_estimators': [50, 100], # specific to random forest; represents number of
#     'classifier__C': [1, 10], # specific fo SVC; controls regularization strength
# }

param_grid_rf = {
    'classifier': [RandomForestClassifier()],
    'classifier__n_estimators': [50, 100],
    'classifier__max_features': ["sqrt", "log2"],
}

param_grid_svc = {
    'classifier': [SVC()],
    'classifier__C': [1, 5],  # Regularization parameter
}

param_grid_lr = {
    'classifier': [LogisticRegression()],
    'classifier__C': [1, 5],  # Regularization parameter
}

param_grid_gb = {
    'classifier': [GradientBoostingClassifier()],
    'classifier__n_estimators': [50, 100],
    'classifier__learning_rate': [0.5], 
    'classifier__max_features': ["sqrt", "log2"], 
}

# Combine the param grids into a list
all_param_grids = [param_grid_rf, param_grid_svc, param_grid_lr, param_grid_gb]


# classifiers = {
#     'GradientBoosting': (GradientBoostingClassifier(), {'n_estimators': [100, 150, 200], 'learning_rate': [0.5, 0.9], 'max_features': ['sqrt']}),
#     'SVM': (SVC(), {'C': [1, 5, 10, 15], 'kernel': ['rbf', 'linear']}),
#     'LogisticRegression': (LogisticRegression(), {'C': [1, 5, 10, 15]}),
# }


## Do grid search CV for different parameter sets

def GridSearchCV_on_featureset(X_train, featureset_id, all_param_grids=all_param_grids, pipeline=pipeline, scoring=scoring, path_results=path_results):
    print(f"RandomizedSearchCV on {featureset_id}")
    
    # fit the model and assess
    best_params_list = []
    best_estimator_list = []
    cv_result_list = []
    cv_result_list_df = []
    cv_result_list_filtered = []

    for param_grid in all_param_grids:
        grid_search = RandomizedSearchCV(pipeline, param_grid, cv=5, scoring=scoring, refit='accuracy') #GridSearchCV
        grid_search.fit(X_train, y_train)

        # Access the best parameters and best estimator for each classifier
        best_params_list.append( grid_search.best_params_ )
        best_estimator_list.append(grid_search.best_estimator_)

        cv_results = grid_search.cv_results_
        cv_result_list.append(cv_results)

        print(cv_results['mean_test_accuracy'])

        print(f"Best Parameters: {grid_search.best_params_ }")
        print(f"Best Estimator: {grid_search.best_estimator_}")


    for cv_scores in cv_result_list:
        cv_df = pd.DataFrame.from_dict(cv_scores)

        cols_params = [x for x in cv_df.columns if x.startswith("param")]
        cols_scores =  [x for x in cv_df.columns if x.startswith("mean_")]

        cv_result_list_df.append(cv_df)
        cv_result_list_filtered.append(cv_df[cols_params+cols_scores])

    merged_cv_df = pd.concat(cv_result_list_df, join='outer', axis=0)
    merged_cv_df_filtered = pd.concat(cv_result_list_filtered, join='outer', axis=0)

    merged_cv_df.to_csv(os.path.join(path_results, 'merged_cv_df_' + featureset_id+'.csv'))
    merged_cv_df_filtered.to_csv(os.path.join(path_results, 'merged_cv_df_filtered_' + featureset_id + '.csv'))

    return(merged_cv_df)


# 206 minutes to run; about 100 of those on the signif feature one
results_store = []
for feature_set_id, X_train_filtered in X_train_featurefilter.items():
    result = GridSearchCV_on_featureset(X_train_filtered, feature_set_id)
    results_store.append(result)


# # fit the model and assess
# best_params_list = []
# best_estimator_list = []
# cv_result_list = []

# for param_grid in all_param_grids:
#     grid_search = GridSearchCV(pipeline, param_grid, cv=5, scoring=scoring, refit='accuracy')
#     grid_search.fit(X_train, y_train)

#     # Access the best parameters and best estimator for each classifier
#     best_params_list.append( grid_search.best_params_ )
#     best_estimator_list.append(grid_search.best_estimator_)

#     cv_results = grid_search.cv_results_
#     cv_result_list.append(cv_results)

#     print(cv_results['mean_test_accuracy'])

#     print(f"Best Parameters: {grid_search.best_params_ }")
#     print(f"Best Estimator: {grid_search.best_estimator_}")

# cv_result_list_filtered = []
# for cv_scores in cv_result_list:
#     print(cv_scores['params'])
#     cv_df = pd.DataFrame.from_dict(cv_scores)

#     cols_params = [x for x in cv_df.columns if x.startswith("param")]
#     cols_scores =  [x for x in cv_df.columns if x.startswith("mean_")]

#     cv_result_list_filtered.append(cv_df[cols_params+cols_scores])


# Explore featureset performance !! TODO

# how well do different featuresets do given the same architecure? 

# Save relevant data


with open(os.path.join(path_results, 'results_store.pkl'), 'wb') as file:
    pickle.dump(results_store, file)
